requirements.txt
aioconsole==0.7.1
aiofiles==23.2.1
aiohttp==3.9.5
aioshutil==1.5
aiosignal==1.3.1
annotated-types==0.7.0
ansicon==1.89.0
anthropic==0.34.0
anyio==4.4.0
apiclient==1.0.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.1.0
blessed==1.20.0
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
construct==2.10.70
cookiecutter==2.6.0
crawlee==0.1.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
discord==2.3.2
discord.py==2.4.0
distro==1.9.0
docutils==0.21.2
doxypypy==0.8.8.7
editor==1.6.6
eval_type_backport==0.2.0
executing==2.0.1
fake-useragent==1.5.1
fastjsonschema==2.20.0
ffmpeg==1.4
filelock==3.15.4
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.1
grpcio-status==1.62.2
gspread==6.1.2
gspread-formatting==1.2.0
gTTS==2.5.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
iniconfig==2.0.0
inquirer==3.3.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jinxed==1.2.1
jiter==0.5.0
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
langdetect==1.0.9
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.3.0
multidict==6.0.5
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.0
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.36.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
playwright==1.45.1
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydub==0.25.1
pyee==11.1.0
Pygments==2.18.0
pykeepass==4.1.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.1.2
PyQt6==6.7.0
PyQt6-Qt6==6.7.2
PyQt6-sip==13.6.0
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
python-telegram-bot==21.4
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qasync==0.27.1
qtconsole==5.5.2
QtPy==2.4.1
readchar==4.1.0
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
runs==1.2.2
selenium==4.22.0
Send2Trash==1.8.3
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
stack-data==0.6.3
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.3.0
tldextract==5.1.2
tokenizers==0.20.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.6.0
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
widgetsnbextension==4.0.11
wsproto==1.2.0
xmod==1.8.1
yarl==1.9.4
